requirements.txt
timm
torch
torchvision
matplotlib
opencv-python
streamlit
scikit-learn
pillow
numpy